Rename object file to output file
