+++ backend/src/modules/boardgame/boardgame.service.ts
@@ -1,4 +1,4 @@
-import { Boardgame } from './model/boardgame.object';
+import { Boardgame } from './model/boardgame.output';
 import { BoardgameRepository } from './../../infrastructure/repositories/boardgame.repository';
 import { Injectable } from '@nestjs/common';
 import { BoardgameInput } from './dto/boardgame.input';
