+++ backend/src/modules/boardgame/model/boardgame.output.ts
